+++ include/MatrixMath.h
@@ -12,14 +12,14 @@ public:
 	Matrix();
 	~Matrix();
 	Matrix(int row_dim, int col_dim);
-	Matrix(int row_dim, int col_dim, float* elements);
+	Matrix(int row_dim, int col_dim, double* elements);
 	Matrix operator*(const Matrix& m);
-	Matrix operator*(const float& f);
+	Matrix operator*(const double& f);
 	Matrix operator+(const Matrix& m);
 	Matrix Transpose();
 	void CoutMatrix();
 	int Dimensions[2];
-	std::vector<float> Elements;
+	std::vector<double> Elements;
 };
 
 #endif
